Switch to using a more secure crypto library
The original encryption middleware copy-pasted the encryption setup from
the PHP manual as an attempt to a) show how the manual could be used and
b) point out that the code in the manual could/should be improved.
Unfortunately, some people didn't understand that ...

Switch out the homegrown implementation for the PHP Encryption library
from Taylor Hornby.

Fixes #1